udpate customize dialog
